Fix functions used within FAIL_UP frames that return immediately.
Previously return objects could be read, and generator frames removed from,
the wrong continuation.

While I'm here, I can't help but feel that the current mechanism that places
a generator frame on the stack for every function call within a FAIL_UP frame
is inefficient. In:

  for Sys::println(f())

for example, println will never yield values (it always returns immediately),
so plopping a generator frame on the stack for the possibility that it might
yield seems pointless. There again, this might be a sufficiently uncommon
usecase that it's not worth optimising for.