Change the way dynamic type checking is done so that it's less messy in RPython.
RPython's inference is more static than I originally appreciated (it can't
link up a type passed as a function argument with a returned object), which
is perfectly reasonable. That meant the original VM.type_check function was
rather more unwieldy in use than was desireable. This patch removes that
function and replaces it with Builtin.type_check_X functions, which is a bit
messier in one sense, but leads to easier to write code.