A basic Earley parser.
This is not yet complete, and may well be buggy, but it does at least
somewhat work. It implements an extension of the Earley parsing algorithm
described in:

   Recognition is not parsing - SPPF-style parsing from cubic recognisers
   Elizabeth Scott, Adrian Johnstone
   Science of Computer Programming 75 (2010) 55-70

In particular it extends this with the Kleene star ()* and optional ()?
operators.

It also - slightly irritatingly - has to implement a slightly flawed version
of disambiguation to maintain compatibility with old grammars, which relied
on a weird version of this in the old VM's parser. This version of the parser
does more disambiguations correctly, and eventually this compatibility hack
should be able to be removed.