Sort out bootstrapping issues.
Previously, objects of a class X (e.g. Func) could be created before X had a
value assigned to it; that object then appeared to be an instance of Object.
This is mostly harmless, unless something needs to look at the type of that
object (e.g. the Exception / Backtrace dance) at which point confusion
reigned.

This patch creates the shells of the builtin classes up front, and then
populates them later. This should ensure any references to them are correct.