Reduce the cost of constants for the JIT.
This replaces the previous cute (but non-JIT friendly) method with a simpler
approach that means the CONST_GET bytecode is easily elidable (and thus has
near zero cost). Overall, this leads to a roughly 3-5% speed increase when
the JIT is in operation.

This requires a change to the bytecode and thus to the bootstrap compiler and
linker. All previous code will need to be recompiled.