Optimise slot lookups for the JIT.
Having finally got around the final part of what Carl Friedrich was
suggesting in
<http://morepypy.blogspot.com/2011/03/controlling-tracing-of-interpreter-with_21.html>,
I've implemented a similar scheme for Converge. The assumption is that
classes rarely change their fields (most classes never change their fields
after their initial creation), so that in general we can simply elide field
lookups entirely. When a field is changed, all subsequent field lookups on
that class and its subclasses need to be regenerated. This gives a fairly
hefty speed-up to a lot of code, because the trace optimiser can now do a
much better job.