Fix Linux build issues.
To my surprise, there appears to be a cross-platform difference in the way
that arguments are passed to GCC. On some platforms, (I assume) a string is
passed; on others a list to execv. The problem is in the latter. Arguments
like "-lz -lm" are treated as a single argument - with consequent hilarity.