Inline functions when tracing.
Now that RPython / PyPy issue #978 has been fixed (thanks to sterling work by
Armin Rigo and Carl Friedrich Bolz), tracing into functions works correctly.
It's not a win 100% of the time, but it's a win enough of the time to make it
worth enabling and in some cases it's a *huge* win.